==> 04_data_analysis.ipynb <==
## <CENTER><U> 1. IMPORTING BASIC NECESSARY PACKAGES</U></CENTER>

pip install pandas_profiling

import pandas as pd
import seaborn as sns
from matplotlib import pyplot as plt
import pandas_profiling

## <CENTER><U> 2. READING INPUT DATA </CENTER></U>

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

data = pd.read_excel("Data processing/feature_extracted.xlsx")

display(data.head())

## <CENTER><U> 3. DATA ANALYSIS </CENTER></U>

data = data.drop(['Travel Count'],axis=1).reset_index(drop=True)

profile = data.profile_report(title='Data Analysis')
profile.to_file(output_file="Data processing/Data Analysis.html")

print(data.corr())

plt.figure(figsize=(12,12))
sns.heatmap(data.corr(), square=True, cmap='Purples')
sns.set(font_scale=1.5)
plt.rc('font', family='Times New Roman', size=5)
plt.show()
